Update white-label scripts (#1472)
* Update white-label scripts

Question from Slack community said:

In the docs, you recommend to use the sed tool. But I'm not sure of all the places that need changing. It would be great if you can include a section in the docs with specific instructions on how to change bundle/application IDs. 

the answer is here: https://github.com/codemagic-ci-cd/white-label-demo-project so I thought to just add this info to the top of this page so people can self-serve.

* Update content/knowledge-white-label/white-label-scripts.md

Co-authored-by: Taha Tesser <tessertaha@gmail.com>

---------

Co-authored-by: Taha Tesser <tessertaha@gmail.com>
Co-authored-by: nevercode-kevin <68225534+nevercode-kevin@users.noreply.github.com>